feat: refresh passkeys on change
